requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
black==21.9b0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
discord.py==1.7.3
idna==3.3
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.1
regex==2021.10.23
requests==2.26.0
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0